requirements.txt
# File: gurbani-insight/requirements.txt

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Vector database
chromadb==0.4.18

# Data processing
pandas==2.1.1
pypdf==3.17.1
tqdm==4.66.1

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Utility libraries
pydantic==2.4.2